Work around gcc 7-9 bug with deducing lambda return type in discarded constexpr if.
As per fmtlib #1455, gcc 7-9 incorrectly deduces the return type of a lambda as 'void' if it is present in the discarded branch of a constexpr if.

As we use the same format string in both branches we can avoid this before gcc 10 by moving the definition of the format string up.