Work around gcc 7-9 bug FMT_STRING bug.
NanoLog is itself invoked from within constexpr if statements, so the same issue as fmtlib #1455 still occurs.

However as these constexpr if statements are all checking the constexpr _nn_has_format_specifiers() function, and we don't have anything that actually needs to be discarded in the othe branch, we can use a normal if statement and rely on trivial elimination of a compile-time true/false test to invoke the appropriate path.