Fix spurious null-appending in NString::AppendSmall.
bytesOther is used both for appending two small strings into a small string, and for appending two small strings into a large string.

The null terminator only needs to be included when appending into a small string so make that explicit rather than adding it to bytesOther in both cases.