Update NanoLog / NFormat behaviour for gcc / clang quirks.
o gcc 9.x or older cannot deduce the return type of a FMT_STRING within a constexpr, so neither NanoLog nor the the macros that call it can use an if constexpr.

However clang identifies that these can be evaluated at compile time in both debug and release builds, gcc identifies them as constexpr in release builds, and so the only impact is on gcc debug builds (where we will perform two checks for format specifiers).

o As we can't use constexpr if in NFormat clang performs printf-checking on the non-format branch.

This is the case even when it can determine at compile time that this branch will never be taken.

This means we must suppress -Wnon-pod-varargs, as per MSVC, as NanoLogFormat can be passed formattable objects whereas NanoLogPrintf can only accept printf arguments.

o gcc's issue with #pragma within a C++ macro remains, so we need to re-enable that.