Improve NStringScanner::FindPattern when searching non-UTF8 text.
PCRE requires UTF8 input so our search text must be converted to this form prior to searching.

Avoid unnecessary transcoding if the input is already UTF8, and avoid unnecessary byte->codepoint lookups if the input is fixed-width UTF8.

Also apply the offset of the search range to the search results, for searching a substring of the original text.