Use std::shared_ptr to access NRunLoop instances.
Allows the thread-specific NRunLoop created inside an NThread to persist after the thread finishes.

The thread_local reference is destroyed during the thread's static destruction, but a shared pointer allows the runloop to be referenced by the NThread object as well. This lets it be safely stopped by NThread::WaitForCompletion even if the thread has exited.