Maintain lists of custom language / warning levels.
Each per-target list is normalised to real paths to allow simpler matching.

Modifying the languge or warning level of files has generator-specific implications for precompiled header support, so implement workarounds for bugs in both CMake and MSVC to allow both language and warning level to be modified on a per-file basis.