Standardise variable-type object interfaces.
All variable typed objects can now the same set of scalar and object-based types.

Type conversion is prohibited such that a type may only be requested if the value is known to be of that type. The exception is NDictionary, which returns a 0/empty object if no value is present for a key.